% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesROE.R
\name{bayesROE}
\alias{bayesROE}
\title{Bayesian Region of Evidence}
\usage{
bayesROE(
  ee,
  se,
  delta = 0,
  alpha = 0.025,
  meanLim = c(pmin(2 * ee, 0), pmax(0, 2 * ee)),
  sdLim = c(0, 3 * se),
  nGrid = 500,
  relative = TRUE
)
}
\arguments{
\item{ee}{Effect estimate.}

\item{se}{Standard error of effect estimate.}

\item{delta}{Minimally relevant effect size. Defaults to zero. Can also be a
vector of numerical values to separate different regions.}

\item{alpha}{Posterior probability that the effect size is less extreme than
delta (i.e. smaller than delta for positive effect estimates, and larger
than delta for negative effect estimates). Defaults to 0.025.}

\item{meanLim}{Limits of prior mean axis. Defaults to interval between zero
and two times the effect estimate.}

\item{sdLim}{Limits of prior standard deviation axis. Defaults to interval
between zero and three times the standard error.}

\item{nGrid}{Number of grid points (on the standard error axis). Defaults to
500.}

\item{relative}{Logical indicating whether a second x-axis and y-axis with
relative prior mean and relative prior variance should be displayed.
Defaults to TRUE.}
}
\value{
A bayesROE object (a list containing the ggplot object, the data for
    the plot, and the tipping point function)
}
\description{
This function computes and visualizes the Bayesian Region of
    Evidence, i.e. the set of normal priors for an effect size which - when
    combined with the observed data - lead to a specified posterior
    probability for the effect size being more extreme than a specified
    minimally relevant effect size.
}
\examples{
## data with p < 0.025 for H0: delta < 0, but p > 0.025 for H0: delta < 0.3
d <- 0.4
d_se <- 0.1
delta <- c(0, 0.3)
bayesROE(ee = d, se = d_se, delta = delta, meanLim = c(-1, 1))

## reproducing Figure 3 from Höfler (2021)
ee <- 9
se <- 3.9
delta <- c(0, 3.75)
bayesROE(ee = ee, se = se, delta = delta, meanLim = c(-5, 10),
         sdLim = c(0, 12), alpha = 0.05)$plot +
  ggplot2::coord_flip(xlim = c(0, 12), ylim = c(-5, 10))

}
\references{
Pawel, S., Matthews, R. and Held, L. (2021). Comment on
    "Bayesian additional evidence for decision making under small sample uncertainty".
    Manuscript submitted for publication. Code available at
    \url{https://osf.io/ymx92/}

 Höfler, M. (2021, October 28). Bayesian regions of evidence (for normal
 distributions). \doi{10.31234/osf.io/mg23h}
}
\author{
Samuel Pawel
}
