% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesROE.R
\docType{package}
\name{bayesROEpkg}
\alias{bayesROEpkg}
\title{bayesROE: Bayesian Regions of Evidence package}
\description{
The bayesROE package enables computation and visualization of
    the Bayesian Regions of Evidence with the function
    \code{\link{bayesROE}}.
}
\references{
Pawel, S., Matthews, R. and Held, L. (2021). Comment on
    "Bayesian additional evidence for decision making under small sample uncertainty".
    Manuscript submitted for publication. Code available at
    \url{https://osf.io/ymx92/}

Höfler, M. (2021, October 28). Bayesian regions of evidence (for normal
 distributions). \doi{10.31234/osf.io/mg23h}
}
