% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesROE.R
\name{print.bayesROE}
\alias{print.bayesROE}
\title{Print method for bayesROE object}
\usage{
\method{print}{bayesROE}(x, ...)
}
\arguments{
\item{x}{A bayesROE object}

\item{...}{Other arguments}
}
\description{
Print method for bayesROE object
}
